divide tests by lib so there's context
